+++ R/wrAdd.R
@@ -6,6 +6,10 @@
 #' 
 #' The species names in \code{species} must match the spelling and capitalization of \code{species} in \code{\link{WSlit}}. Use \code{wsVal()} to see a list of all species for which standard weight equations exist in \code{\link{WSlit}} and, more importantly, how the species names are spelled and capitalized.
 #' 
+#' The \code{thesaurus} argument may be used to relate alternate species names to the species names used in \code{WSlit}. For example, you (or your data) may use \dQuote{Bluegill Sunfish}, but \dQuote{Bluegill} is used in \code{WSlit}. The alternate species name can be used here if it is defined in a named vector (or list) given to \code{thesarus=}. The alternate species name is the value and the species name in \code{PSDlit} is the name in this vector/list - e.g., \code{c("Bluegill"="Bluegill Sunfish")}. See the examples for a demonstration.
+#' 
+#' Some species have length categories separated by sub-group. For example, length categories exist for both lentic and lotic populations of Brown Trout. The length values for a sub-group may be obtained by either including the species name in \code{species} and the sub-group name in \code{group} in \code{WsOpts} or by using the combined species and sub-group name, with the sub-group name in parentheses, in \code{species}. Both methods are demonstrated in the examples. Note that an error is returned if a species has sub-groups but neither method is used to define the sub-group.
+#' 
 #' Some (few) species have more than one equation listed in \code{\link{WSlit}} (for the specified units). In these instances the user must select one of the equations to use with \code{WsOpts}. \code{WsOpts} is a list of lists where the inside list contains one or more of \code{group}, \code{ref}, or \code{method} (see \code{\link{WSlit}}) required to specify a single equation for a particular species, which is the name of the inner list. See the examples for an illustration of how to use \code{WsOpts}.
 #' 
 #' See examples and \href{https://fishr-core-team.github.io/FSA/articles/Computing_Relative_Weights.html}{this article} for a demonstration.
@@ -244,3 +248,42 @@ wrAdd.formula <- function(wt,data,thesaurus=NULL,units=c("metric","English"),...
   wrAdd.default(tmp$mf[,tmp$Rpos],tmp$mf[,tmp$ENumPos],
                 tmp$mf[,tmp$EFactPos],thesaurus,units,...)
 }
+
+# ==============================================================================
+# Internal -- prepare WSlit by loading it and replacing its default names with
+#             names in the thesaurus, if any
+# ==============================================================================
+iPrepWSlit <- function(thesaurus) {
+  # Load WSlit into dat in this function namespace
+  dat <- FSA::WSlit
+  if (!is.null(thesaurus)) {
+    # Some sanity checks on thesaurus
+    if (!(is.vector(thesaurus) | is.list(thesaurus)))
+      STOP("'thesaurus' must be either a vector or list. ",
+           "Make sure it is not 'factor'ed.")
+    if (length(names(thesaurus))==0)
+      STOP("Values in 'thesaurus' must be named (with species names from 'WSlit'.)")
+    if (!is.character(thesaurus))
+      STOP("Values in 'thesaurus' must be strings of species names.")
+    # thesaurus appears to be a named vector/list of strings ... start processing
+    # Alphabetize names in thesaurus to match dat/WSlit
+    thesaurus <- thesaurus[order(names(thesaurus))]
+    # Remove name from thesaurus if not in dat/WSlit
+    thes.nokeep <- which(!names(thesaurus) %in% unique(dat$species))
+    if (length(thes.nokeep)>0) {
+      MESSAGE("The following species names were in 'thesaurus' but do ",
+              "not have an entry in 'WSlit' and will be ignored: ",
+              iStrCollapse(names(thesaurus)[thes.nokeep]),".")
+      thesaurus <- thesaurus[-thes.nokeep]      
+    }
+    # Find species in dat/PSDlit in kept thesaurus and change names to thesaurus names
+    if (length(thesaurus)>0) {
+      tmp <- match(dat$species,names(thesaurus))
+      thes.pos <- which(!is.na(tmp))
+      dat$species[thes.pos] <- unname(thesaurus[tmp[thes.pos]])
+    }
+  }
+  # Return dat/WSlit
+  dat
+}
+